Added files for the lorenz butterfly project
Contains the Lorenz Butterfly object that generates the dataset used to train the models and demonstrates my skills in Object Orientated Programming in python. Along with the notebook where I generated visuals of the butterfly and created the two models I chose to model and predict the butterly along with analytical and communicative aspects that drove comparison and derived possible improvements for design and building of the models.